% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center}
\usage{
center(data, midpoint = 0)
}
\arguments{
\item{data}{A vector of data to transform}

\item{midpoint}{A value around which to center the data}
}
\value{
A vector of equal length
}
\description{
A function that centers the data around a chosen midpoint
}
\details{
This function takes a vector of data and retrurns a vector of equal length with the vlaues trnasformed to be centered around a new value. By defualt this value is 0
}
