% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\name{europe_07}
\alias{europe_07}
\title{European 2007 Data}
\usage{
europe_07(data, plot = FALSE, denomination = "B")
}
\arguments{
\item{data}{The raw dataset}

\item{plot}{This will generate a plot of GDP vs life expectancy}

\item{denomination}{The value in which to give GDP, excepts T = Trillion, B = Billion or M= Millions}
}
\value{
returns a dataset
}
\description{
A function that converts the raw dataset into the need 2007 data.
}
\details{
This datset is used in a pipeline which takes in the raw datset, and produces a GDP value, and gives this value in a denomination of choice, either Trillion, Billion or Million.
}
